feat(cli): split into cobra subcommands; foundation for v0.7 DX
Introduces cobra and a new canonical entrypoint at cmd/duragraph. The
existing cmd/server is reduced to a 4-line backwards-compat shim so
deploy/docker/Dockerfile.{api,server} (which build ./cmd/server) keep
producing a working binary.

New subcommand layout (see duragraph-spec/backend/binary-modes.yml §
subcommands):

  duragraph                # backwards-compat: same as `serve`
  duragraph serve          # fully implemented (lifted from cmd/server)
  duragraph version        # implemented; --json for machine output
  duragraph dev            # stub — phases 2-5 of binary-modes.yml
  duragraph init           # stub — phase 6
  duragraph migrate        # stub — explicit migration control
  duragraph runs {tail,get,trigger}  # stubs — phase 7
  duragraph events tail    # stub — phase 7
  duragraph studio {install,uninstall}  # stubs — phase 8

Stubs all return a "not yet implemented — see binary-modes.yml" error
and exit 1, so follow-up PRs only have to fill in the body.

Mechanics of the serve move:
  - cmd/server/main.go's main() body is now serve.go's runServe(); the
    only diffs vs the original are (a) func signature wrapped as cobra
    RunE and (b) log.Fatalf(...) → return fmt.Errorf(...). The latter
    is a strict improvement: the original log.Fatalf skipped defers, so
    pgxpool / NATS connection close was leaked on startup-time errors;
    return-error-from-RunE runs the defers.
  - cmd/server/version.go and cmd/server/metrics_bootstrap.go moved to
    cmd/duragraph/cmd/ unchanged (package rename only).
  - The 700+ line serve body is intentionally NOT refactored in this
    PR — that's deferred to phase 2/3 when TOML config + flags arrive.

No new env vars, no new flags, no spec changes. Existing deployments
keep working unchanged: cmd/server still builds, the binary still
honours all the existing env vars (DB_HOST, NATS_URL, AUTH_ENABLED,
MIGRATOR_PLATFORM_ENABLED, …), and a bare invocation defaults to
serve.

Verified:
  - go build ./...        clean
  - go vet ./...          clean
  - go test ./... -short  all pass
  - go build ./cmd/server (shim) produces a working binary
  - go build ./cmd/duragraph produces a working binary
  - duragraph version / version --json / --help / dev (exits 1) all
    behave per spec
